ugc/requirements.txt
kafka-python==2.0.2
Flask==2.2.2
flask-jwt-extended==4.4.4
cryptography==38.0.4
pydantic==1.9.0
flake8==5.0.4
PyYAML==6.0
gevent==22.10.2
gunicorn==20.1.0
python-dotenv==0.21.0
backoff==2.2.1
